fix: adiciona tratamento de exceções aos métodos

+++ src/main/java/com/example/ControlHabits/service/HabitService.java
@@ -3,6 +3,7 @@ package com.example.ControlHabits.service;
 import com.example.ControlHabits.dto.CreateHabitDTO;
 import com.example.ControlHabits.dto.UpdateHabitDTO;
 import com.example.ControlHabits.entity.Habit;
+import com.example.ControlHabits.exception.HabitNotFoundException;
 import com.example.ControlHabits.repository.HabitRepository;
 import org.springframework.stereotype.Service;
 
@@ -17,37 +18,38 @@ public class HabitService {
         this.habitRepository = habitRepository;
     }
 
-    public Habit createHabits(CreateHabitDTO habitDTO) {
+    public Habit createHabit(CreateHabitDTO habitDTO) {
         Habit habit = habitDTO.toEntity();
         return habitRepository.save(habit);
     }
+
     public List<Habit> getAllHabits() {
         return habitRepository.findAll();
-
     }
 
     public void deleteHabit(Long id) {
+        if (!habitRepository.existsById(id)) {
+            throw new HabitNotFoundException(id);
+        }
         habitRepository.deleteById(id);
     }
 
     public Habit getHabitById(Long id) {
-        return habitRepository.findById(id).orElse(null);
+        return habitRepository.findById(id)
+                .orElseThrow(() -> new HabitNotFoundException(id));
     }
 
     public Habit updateHabit(Long id, UpdateHabitDTO updateHabitDTO) {
+        Habit habit = habitRepository.findById(id)
+                .orElseThrow(() -> new HabitNotFoundException(id));
 
-        Habit habit = habitRepository.findById(id).orElse(null);
-
-        if (habit != null) {
-            if (updateHabitDTO.getName() != null) {
-                habit.setName(updateHabitDTO.getName());
-            }
-            if (updateHabitDTO.getDescription() != null) {
-                habit.setDescription(updateHabitDTO.getDescription());
-            }
-            return habitRepository.save(habit);
-        } else {
-            return null;
+        if (updateHabitDTO.getName() != null) {
+            habit.setName(updateHabitDTO.getName().trim());
         }
+        if (updateHabitDTO.getDescription() != null) {
+            habit.setDescription(updateHabitDTO.getDescription().trim());
+        }
+
+        return habitRepository.save(habit);
     }
 }
